modulo-biometrico/requirements.txt
flask
deepface
tensorflow==2.11
pillow
